:sparkles: feat: Tratamento de exceções

+++ src/main/java/com/java/transacoes_api/conta/services/ContaService.java
@@ -5,6 +5,7 @@ import com.java.transacoes_api.conta.entities.Conta;
 import com.java.transacoes_api.conta.exceptions.ContaNaoEncontradaException;
 import com.java.transacoes_api.conta.exceptions.UsuarioNaoEncontradoException;
 import com.java.transacoes_api.conta.repository.ContaRepository;
+import com.java.transacoes_api.movimentacao.repository.MovimentacaoRepository;
 import com.java.transacoes_api.usuario.repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,6 +21,9 @@ public class ContaService {
 
     @Autowired
     private UsuarioRepository usuarioRepository;
+
+    @Autowired
+    private MovimentacaoRepository movimentacaoRepository;
 
     public Conta criarConta(ContaInputDTO dto)  throws Exception {
         var usuarioExistente = usuarioRepository.findById(dto.usuarioId())
@@ -45,6 +49,12 @@ public class ContaService {
 
         var contaExistente = contaRepository.findById(id)
                 .orElseThrow(ContaNaoEncontradaException::new);
+
+        boolean possuiMovimentacao = movimentacaoRepository.existsByContaId(id);
+
+        if(possuiMovimentacao) {
+            throw new RuntimeException("Não é permitida a exclusão da conta com movimentações realizadas");
+        }
 
         contaRepository.deleteById(id);
     }
